refactored Sidebar into its own component
