+++ client/src/App.tsx
@@ -1,4 +1,4 @@
-import { useContext, useEffect, useState } from 'react';
+import { useEffect, useState } from 'react';
 import { styled } from '@mui/material/styles';
 import { Route, Routes } from 'react-router-dom';
 import Home from './routes/Home';
@@ -6,26 +6,17 @@ import Bubbles from './routes/Bubbles';
 import './App.css';
 import Header from './components/Header';
 import NewsFeed from './components/NewsFeed';
-import { UserContext, UserProvider } from './contexts/UserContext';
+import { UserProvider } from './contexts/UserContext';
 import { CoinProvider } from './contexts/CoinContext';
 import { Coin } from './types/Coin';
 import PasswordReset from './routes/PasswordReset';
 import { Box } from '@mui/material';
-import Drawer from '@mui/material/Drawer';
 // import CssBaseline from '@mui/material/CssBaseline';
 import MuiAppBar, { AppBarProps as MuiAppBarProps } from '@mui/material/AppBar';
 import Toolbar from '@mui/material/Toolbar';
-import List from '@mui/material/List';
 import IconButton from '@mui/material/IconButton';
 import MenuIcon from '@mui/icons-material/Menu';
-import ChevronLeftIcon from '@mui/icons-material/ChevronLeft';
-import ListItem from '@mui/material/ListItem';
-import ListItemButton from '@mui/material/ListItemButton';
-import ListItemIcon from '@mui/material/ListItemIcon';
-import ListItemText from '@mui/material/ListItemText';
-import InboxIcon from '@mui/icons-material/MoveToInbox';
-import MailIcon from '@mui/icons-material/Mail';
-import AuthDialog from './components/AuthDialog';
+import Sidebar from './components/Sidebar';
 
 // const DynamicLoader = ({ component }: { component: string }) => {
 //   const LazyComponent = useMemo(
@@ -60,15 +51,6 @@ const Main = styled('main', { shouldForwardProp: (prop) => prop !== 'open' })<{
   }),
 }));
 
-const DrawerHeader = styled('div')(({ theme }) => ({
-  display: 'flex',
-  alignItems: 'center',
-  padding: theme.spacing(0, 1),
-  // necessary for content to be below app bar
-  ...theme.mixins.toolbar,
-  justifyContent: 'flex-end',
-}));
-
 interface AppBarProps extends MuiAppBarProps {
   open?: boolean;
 }
@@ -95,7 +77,6 @@ function App() {
   const [globalData, setGlobalData] = useState({});
   const [searchText, setSearchText] = useState<string>('');
   const [isOpen, setIsOpen] = useState(false);
-  const userContext = useContext(UserContext);
 
   useEffect(() => {
     fetchGlobal();
@@ -114,81 +95,6 @@ function App() {
       .then((res) => res.json())
       .then((res) => setCoins(res.data))
       .catch((err) => console.error('Error', err));
-  };
-
-  const [isLoginDialogOpen, setIsLoginDialogOpen] = useState<boolean>(false);
-  const [isSignupDialogOpen, setIsSignupDialogOpen] = useState<boolean>(false);
-
-  const MenuItems = [
-    {
-      name: 'Login',
-      handleClick: () => setIsLoginDialogOpen(true),
-    },
-    {
-      name: 'Signup',
-      handleClick: () => setIsSignupDialogOpen(true),
-    },
-  ];
-
-  const handleLoginDialogSubmit = ({
-    email,
-    password,
-  }: {
-    email: string;
-    password: string;
-  }) => {
-    handleLogin({ email, password });
-    setIsLoginDialogOpen(false);
-  };
-
-  const handleLogin = async ({
-    email,
-    password,
-  }: {
-    email: string;
-    password: string;
-  }) => {
-    const res = await fetch('/api/login', {
-      method: 'POST',
-      headers: {
-        'Content-Type': 'application/json',
-      },
-      body: JSON.stringify({ email, password }),
-    });
-
-    const data = await res.json();
-    localStorage.setItem('user', data.data.user.session.access_token);
-    userContext.setUser(data.data.user.session.access_token);
-    const newFavorites = data.data.favorites.map((c: Coin) => c.name);
-    userContext.setFavoriteCoins(newFavorites);
-  };
-
-  const handleSignupDialogSubmit = ({
-    email,
-    password,
-  }: {
-    email: string;
-    password: string;
-  }) => {
-    handleSignup({ email, password });
-    setIsSignupDialogOpen(false);
-  };
-
-  const handleSignup = async ({
-    email,
-    password,
-  }: {
-    email: string;
-    password: string;
-  }) => {
-    await fetch('/api/signup', {
-      method: 'POST',
-      headers: {
-        'Content-Type': 'application/json',
-      },
-      body: JSON.stringify({ email, password }),
-    });
-    // TODO: check res and tell user to check email or login
   };
 
   return (
@@ -214,62 +120,11 @@ function App() {
               />
             </Toolbar>
           </AppBar>
-          <Drawer
-            sx={{
-              width: drawerWidth,
-              flexShrink: 0,
-              '& .MuiDrawer-paper': {
-                width: drawerWidth,
-                boxSizing: 'border-box',
-                background: (theme) => theme.palette.primary.main,
-                display: 'flex',
-                justifyContent: 'space-between',
-              },
-            }}
-            variant="persistent"
-            anchor="left"
-            open={isOpen}
-          >
-            <DrawerHeader>
-              <IconButton onClick={() => setIsOpen(false)}>
-                <ChevronLeftIcon />
-                {/* {theme.direction === 'ltr' ? (
-                  
-                ) : (
-                  <ChevronRightIcon />
-                )} */}
-              </IconButton>
-            </DrawerHeader>
-            <List>
-              {['Notifications'].map((text, index) => (
-                <ListItem key={text} disablePadding>
-                  <ListItemButton>
-                    <ListItemIcon>
-                      {index % 2 === 0 ? <InboxIcon /> : <MailIcon />}
-                    </ListItemIcon>
-                    <ListItemText primary={text} />
-                  </ListItemButton>
-                </ListItem>
-              ))}
-            </List>
-            <List>
-              {MenuItems.map((menuItem) => (
-                <ListItem key={menuItem.name} disablePadding>
-                  <ListItemButton
-                    onClick={() => {
-                      menuItem.handleClick();
-                    }}
-                    alignItems="center"
-                  >
-                    <ListItemText
-                      primary={menuItem.name}
-                      sx={{ display: 'flex', justifyContent: 'center' }}
-                    />
-                  </ListItemButton>
-                </ListItem>
-              ))}
-            </List>
-          </Drawer>
+          <Sidebar
+            isOpen={isOpen}
+            setIsOpen={setIsOpen}
+            drawerWidth={drawerWidth}
+          />
           <Main open={isOpen}>
             <Box display="flex" flexDirection="column" mt={3}>
               <NewsFeed />
@@ -281,18 +136,6 @@ function App() {
             </Box>
           </Main>
         </Box>
-        <AuthDialog
-          open={isLoginDialogOpen}
-          onClose={() => setIsLoginDialogOpen(false)}
-          onSubmit={handleLoginDialogSubmit}
-          mode="login"
-        />
-        <AuthDialog
-          open={isSignupDialogOpen}
-          onClose={() => setIsSignupDialogOpen(false)}
-          onSubmit={handleSignupDialogSubmit}
-          mode="signup"
-        />
       </CoinProvider>
     </UserProvider>
   );
